KL-151/chore: move oauth files into oauth
