+++ src/main/java/taco/klkl/global/error/exception/ErrorCode.java
@@ -75,10 +75,12 @@ public enum ErrorCode {
 	TOKEN_EXPIRED(HttpStatus.UNAUTHORIZED, "만료된 토큰입니다."),
 	TOKEN_GENERATION_FAILED(HttpStatus.INTERNAL_SERVER_ERROR, "토큰 생성 중 오류가 발생했습니다."),
 
+	// OAuth
+	REGISTRATION_ID_NOT_FOUND(HttpStatus.NOT_FOUND, "유효하지 않은 접속 ID 입니다."),
+
 	// Authentication
 	UNAUTHORIZED(HttpStatus.UNAUTHORIZED, "인증되지 않은 사용자입니다."),
 	PERMISSION_DENIED(HttpStatus.FORBIDDEN, "권한이 거부되었습니다."),
-	REGISTRATION_ID_NOT_FOUND(HttpStatus.NOT_FOUND, "유효하지 않은 접속 ID 입니다."),
 
 	// Sample
 	SAMPLE_ERROR(HttpStatus.BAD_REQUEST, "샘플 에러입니다."),
